:recycle: change authService  and remove unnecessary repository
