--- Server/U-Tracker API/src/main/kotlin/com/code_of_duty/utracker_api/data/dao/LoginDao.kt
@@ -1,7 +1,0 @@
-package com.code_of_duty.utracker_api.data.dao
-
-import com.code_of_duty.utracker_api.data.models.Student
-
-interface LoginDao {
-    fun findByCode(code: String?): Student?
-}
