+++ Server/U-Tracker API/src/main/kotlin/com/code_of_duty/utracker_api/services/auth/AuthService.kt
@@ -1,44 +1,12 @@
 package com.code_of_duty.utracker_api.services.auth
 
-import com.code_of_duty.utracker_api.data.dao.LoginDao
-import com.code_of_duty.utracker_api.data.dao.StudentDao
 import com.code_of_duty.utracker_api.data.dtos.RegisterDto
 import com.code_of_duty.utracker_api.data.models.Student
-import com.code_of_duty.utracker_api.utils.PasswordUtils
 import org.springframework.stereotype.Service
 
 @Service
-class AuthService(private val studentDao: StudentDao,
-                  private val loginDao: LoginDao,
-                  private val passwordUtils: PasswordUtils) {
-
-    fun isCodeTaken(code: String): Boolean {
-        return studentDao.existsByCode(code)
-    }
-
-    fun registerStudent(registerDto: RegisterDto, degree: String): Student {
-        if (isCodeTaken(registerDto.code)) {
-            throw IllegalArgumentException("Code already taken")
-        }
-
-        if (registerDto.password != registerDto.confirmPassword) {
-            throw IllegalArgumentException("Passwords do not match")
-        }
-
-        val hashPassword = passwordUtils.hashPassword(registerDto.password)
-        val newStudent = Student(code = registerDto.code, username = registerDto.username, hashPassword = hashPassword)
-        return studentDao.save(newStudent)
-    }
-
-    fun authenticate(code: String, password: String): Student? {
-        if (!studentDao.existsByCode(code)) {
-            return null
-        }
-        val student = loginDao.findByCode(code)
-        if (student == null || !passwordUtils.verifyPassword(password, student.hashPassword)) {
-            return null
-        }
-        return student
-    }
-
+interface AuthService {
+    fun registerStudent(registerDto: RegisterDto, degree: String): Student
+    fun authenticate(code: String, password: String): Student?
+    fun isCodeTaken(code: String): Boolean
 }
