+++ Server/U-Tracker API/src/main/kotlin/com/code_of_duty/utracker_api/services/auth/AuthServiceImp.kt
@@ -1,23 +1,40 @@
 package com.code_of_duty.utracker_api.services.auth
 
-import com.code_of_duty.utracker_api.data.dao.LoginDao
+import com.code_of_duty.utracker_api.data.dao.StudentDao
+import com.code_of_duty.utracker_api.data.dtos.RegisterDto
 import com.code_of_duty.utracker_api.data.models.Student
-import jakarta.persistence.EntityManager
-import jakarta.persistence.NoResultException
-import org.springframework.stereotype.Repository
+import com.code_of_duty.utracker_api.utils.PasswordUtils
+import org.springframework.stereotype.Component
 
-@Repository
-class AuthServiceImp(private val entityManager: EntityManager) : LoginDao {
-    override fun findByCode(code: String?): Student? {
-        val query = entityManager.createQuery(
-            "SELECT s FROM Student s WHERE s.code = :code",
-            Student::class.java
-        )
-        query.setParameter("code", code)
-        return try {
-            query.singleResult
-        } catch (e: NoResultException) {
-            null
+@Component
+class AuthServiceImp(private val studentDao: StudentDao,
+                     private val passwordUtils: PasswordUtils): AuthService{
+
+    override fun isCodeTaken(code: String) = studentDao.existsByCode(code)
+
+    override fun registerStudent(registerDto: RegisterDto, degree: String): Student {
+        if (isCodeTaken(registerDto.code)) {
+            throw IllegalArgumentException("Code already taken")
         }
+
+        if (registerDto.password != registerDto.confirmPassword) {
+            throw IllegalArgumentException("Passwords do not match")
+        }
+
+        val hashPassword = passwordUtils.hashPassword(registerDto.password)
+        val newStudent = Student(code = registerDto.code, username = registerDto.username, hashPassword = hashPassword)
+        return studentDao.save(newStudent)
     }
+
+    override fun authenticate(code: String, password: String): Student? {
+        if (!studentDao.existsByCode(code)) {
+            return null
+        }
+        val student = studentDao.findById(code).orElse(null)
+        if (student == null || !passwordUtils.verifyPassword(password, student.hashPassword)) {
+            return null
+        }
+        return student
+    }
+
 }
