--- Server/U-Tracker API/src/main/kotlin/com/code_of_duty/utracker_api/services/student/StudentImp.kt
@@ -1,4 +1,0 @@
-package com.code_of_duty.utracker_api.services.student
-
-class StudentImp {
-}
